+++ src/prompt-templates/src/handlebars/index.ts
@@ -1,0 +1,1 @@
+export * from './handlebarsPromptTemplate';
